local-agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "local-agents"
version = "0.1.0"
description = "Local Strands agents (Claude, Ollama)"
requires-python = ">=3.10"
dependencies = [
    "strands-shared",
    "bedrock-agentcore>=1.0.6",
    "strands-agents>=1.16.0",
    "strands-agents-tools>=0.2.15",
    "boto3>=1.40.0",
    "botocore>=1.40.0",
    "mcp>=1.0.0",
    "playwright>=1.56.0",
    "opencv-python>=4.12.0",
    "pytesseract>=0.3.13",
    "numpy>=2.0.0,<2.3.0",
    "pillow>=11.2.1,<12.0.0",
    "ollama>=0.6.1",
    "rich>=13.0.0",
]

[project.scripts]
local-agents = "local_agents.cli:main"

[tool.uv.sources]
strands-shared = { workspace = true }
